Clarify and distinguish usage and error messages.
Tcpslice used to print the version and usage message as an error message
to stderr.

Introduce a "-h" option to print the message to stdout as that is the
normal expected result of it, not an error condition. In case of an
actual error prepend the message on stderr with an explicit error line.
Print libpcap version next to tcpslice version like tcpdump does.