Don't set CFLAGS to include -Werror when running the configure script.
That causes tests such as the AC_REPLACE_FUNCS() tests to fail, because
that test relies on an old-style function declaration getting a warning,
rather than an error, if, for example, the function being tested is a
built-in.

Instead, pass CFLAGS=-Werror in the make command; that matches what we
do for tcpdump.