[2/3] Add parseable timestamp format syntax testing.
Add functions to test parseable timestamp strings syntax-wise (whether
amounts and valid units are present and follow a correct order), but not
semantics-wise (whether the amounts and the resulting date and time are
valid).

Add a function to test for all known timestamp format variations. To tell
filenames from timestamps after the getopt() loop use that function
instead of testing the first character only.

This reduces the problem space further, this time enough to stop making
mistakes when there is no ambiguity. Update the man page.