Fixup the previous commit.
After additional testing it became apparent that the purpose of the
custom rule was to use version.c from the build directory (where another
rules generates it), not the source directory. So it was my change that
had broken out-of-tree build, not something else. Undo this change and
apply a fix for the original problem.