CI: Refine the build matrix scripts.
If not provided with a pre-arranged prefix path, make a temporary one
and remove it when done. Between the rounds delete the contents of the
prefix directory, but not the directory. Quote the prefix variable
properly. Print OS and compiler versions at the beginning. Do not cat
Makefile twice in the Travis CI block.