CI: Refine the build matrix scripts more.
Fix skipping GCC on macOS in build_matrix.sh to make it work
consistently in and outside of a CI environment. Remove the skipping
from the CI space. Print Makefile in a way that works reliably for all
repositories and both build systems. Require travis_fold() parameters to
be set.