CI: Add MATRIX_BUILD_LIBPCAP to pass on FreeBSD.
The previously implemented approach was to let ./configure detect
libpcap, which would be either a local libpcap (if the caller has
pre-built it in ../libpcap) or a system libpcap (if it has not). In
particular, the Buildbot master setup arranged a local libpcap, which
defaulted to GCC [1]. The Buildbot workers that could compile tcpslice
varied in which libpcap they were able to use:

* freebsd-aarch64: system only [1]
* linux-aarch64: both
* linux-armv7l: both
* linux-ppc64le: both
* netbsd-aarch64: local only
* netbsd-amd64: local only

To simplify the builder commands and to allow customisation at the same
time, copy the MATRIX_BUILD_LIBPCAP logic from tcpdump and use it in
worker-specific environment variables as necessary.

A useful side effect of this change is compiling the local libpcap using
the same CC as for the current round of tcpslice. This works around a
linking issue [1], restores tcpslice-freebsd-aarch64 and allows it and
other builders to test two additional permutations.

1: FreeBSD shortly before release 13.0 started to have an odd issue, in
that tcpslice compiled using Clang failed to link with local libpcap
compiled using GCC:

ld: error: ./pcap.c:736:(.text+0x116C): unknown relocation (313) against
symbol pcap_netmap_findalldevs

ld: error: ./pcap.c:2335:(.text+0x1548): unknown relocation (313)
against symbol pcap_netmap_create

ld: error: ./pcap.c:3502:(.text+0x2210): unknown relocation (313)
against symbol __stderrp

(and so on)