CI: Introduce and use TCPSLICE_TAINTED.
The new environment variable provides means to declare all or some
tcpslice build steps of a specific environment (CI or a working copy) as
tainted. Add explicit conditions to exempt the builds that emit warnings
now and request the compiler to treat warnings as errors *iff* the build
is not tainted. This way if a build that is warning-free now degrades
later, CI will fail it.

Call the variable TCPSLICE_TAINTED so any nested libpcap builds can use
their own logic to manage their own CFLAGS.