Squelch pointer conversion warnings in sessions.c.
Although compilers implement function-to-void pointer conversion, there
seems to be a difference in opinion whether it deserves a warning and
which level and wording would be the most appropriate.  That said,
libnids headers are unlikely to fix their expected pointer type, and the
warnings have been tainting Linux builds that have libnids-dev, so let's
just squelch them.

In compiler-tests.h synchronize Clang and XL C macros with libpcap.
Port diag-control.h from tcpdump, spell XL C tests and pragmas better
and put the default empty definitions after the if block to make it
easier to follow.  Remove the exemption rule.