Require every input file to have the same DLT.
tcpslice uses DLT of the first input file as DLT of the output file.
However, it did not check if any other input file has a different DLT,
so if the types were not the same, the output would likely contain
invalid (as far as its DLT goes) packets, and there would be no reliable
way to recover this data afterwards.

Factor warning() and complain() out of error(). Add and use
validate_files() to eliminate the failure to fail by rejecting all input
if the DLT is not the same in all files.