Allow disabling optional extra libraries.
In our build system, a package may be built in any order, meaning that
when built today libnids may not be present, and when built tomorrow
libnids may be present, just because of someone else's dependency.

In order to create repeatable builds, we disable optional libraries that
we don't explicitly want, so this pull request adds AC_ARG_WITH
statements to add --without-foo for libnids, libosipparser2, libooh323c.