autoconf: Add the option to print functions and files names
This should help some debugging or learning processes.

Usage:
./configure --enable-instrument-functions

Generate instrumentation calls for entry and exit to functions.
Just after function entry and just before function exit, these
profiling functions are called and print the function names with
indentation and call level.
If entering in a function, print also the calling function name with
file name and line number. There may be a small shift in the line number.

To configure the printing of only the global functions names:
$ make instrument_global

To go back to print all the functions names:
$ make instrument_all

To print nothing, like with no instrumentation:
$ make instrument_off

The library libbfd is used, therefore the binutils-dev package is required.

[skip ci]