Get the temporary directory right in "make releasetar".
The Buildbot workers started to run out of disk space because the old
temporary directories were left in place even after successful runs
because I suggested to prepend an additional /tmp to AUTORECONF_DIR,
which was wrong.

Lose the excess prefix given to "rm -f" to fix the problem for succesful
Buildbot builds.  Refine the placement and naming of AUTORECONF_DIR to
reuse existing automatic eventual cleanup for failed Buildbot builds.