Autoconf: Squelch a few more warnings from 2.71.
The minimum required version is 2.69, but 2.71 is becoming quite common
(in particular, in the CI context), so it is useful to keep the number
of generated warnings low.

Replace AC_CONFIG_HEADER with AC_CONFIG_HEADERS, which works the same
with Autoconf 2.69 and squelches a warning with Autoconf 2.71:

configure.ac:156: warning: The macro `AC_CONFIG_HEADER' is obsolete.

Likewise, convert AC_TRY_COMPILE to AC_COMPILE_IFELSE to address a
similar future issue (or maybe not, considering the last release of
Ultrix was 28 years ago):

configure.ac:31: warning: The macro `AC_TRY_COMPILE' is obsolete.

Fix a comment while at it.