Autoconf: Retire Ultrix-specific workarounds.
Remove the workaround from Makefile: if Ultrix make was the only make
where the problem existed, then the problem no longer stands.  If any
other supported OS make still requires the workaround, it can be
reinstated with an up-to-date comment.

Remove the const support detection from aclocal.m4: what it did was
similar or maybe even the same as AC_C_CONST, but as Autoconf 2.69
documentation says: "This macro is obsolescent, as current C compilers
support const.  New programs need not use this macro."  This is
consistent with the fact that currently supported OSes manage to compile
tcpslice without AC_C_CONST.