instrument-functions: leave space for a terminating '\0'.
readlink() just reads the raw symlink data, which doesn't include a
terminating '\0'.  We must leave space for the '\0', and *not* tell
readlink() about that, so that, when we use its return value to tell us
where to put the '\0', it doesn't return a value equal to the full size
of the buffer, meaning we'd put the '\0' *after* the end of the buffer.

Fixes Coverity CID 310987.