Look for the VERSION file in the source directory.
This matches libpcap and tcpdump.

It used to be necessary for out-of-tree builds.  It's no longer needed
for that, but at least it means that attempts to build a release tarball
will fail with "you're not in a git clone directory" rather than "I
can't find the VERSION file".

Now that the configure script is no longer in the repository, but is
generated by autogen.sh, and given that configure.ac gets the package
version from the VERSION file, we should probably just use
PACKAGE_VERSION in Makefile.in rather than `cat $(srcdir)/VERSION`.