Spell BSD-3-Clause in sessions.[ch].
The only differences between the licence in these two files and the
reference wording at [1] are as follows:

* Clause 3 says "The name of the author may not be used..." in the two
  files and "Neither the name of the copyright holder nor the names of
  its contributors may be used..." in the reference wording.

* The first sentence of the disclaimer says "THIS SOFTWARE IS
  PROVIDED..." in the two files and "THIS SOFTWARE IS PROVIDED BY THE
  COPYRIGHT HOLDERS AND CONTRIBUTORS..." in the reference wording.

* The second sentence of the disclaimer says "IN NO EVENT SHALL THE
  AUTHOR BE LIABLE..." in the two files and "IN NO EVENT SHALL THE
  COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE..." in the reference
  wording.

According to [1], all of these are known permitted variants.  To clarify
the matter, switch to the exact reference wording and formatting, this
makes the copyright line, clause 3 and the SPDX licence ID the only
differences compared to BSD-2-Clause files.

1: https://spdx.org/licenses/BSD-3-Clause.html